requirements.txt
Flask==2.3.3
Flask-Dropzone==1.6.0
Flask-Session==0.5.0
Flask-WTF==1.1.1
WTForms==3.0.1
Pillow==10.0.1
opencv-python==4.8.1.78
pytesseract==0.3.10
googletrans==4.0.0rc1
gTTS==2.3.2
numpy==1.24.3
requests==2.31.0